pyproject.toml
[project]
name = "appointment-scheduler-2-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "python-dotenv>=1.1.0",
    "supabase>=2.15.1",
    "uvicorn>=0.34.2",
]
